requirements.txt
# --- Twxn's Prepaid Market dependencies ---
requests==2.32.3
python-dotenv==1.0.1
cryptography==43.0.3
SQLAlchemy==2.0.34
aiogram==3.22.0
aiohttp==3.9.5
pydantic==2.9.2
pydantic-core==2.23.4
aiosqlite==0.20.0
bip-utils==2.9.3
# Optional: for Solana or BTC HD wallet derivations
# solders==0.27.0

